modify merge.c

+++ 2D/MPI/common/merge.c
@@ -170,8 +170,8 @@ int main(int argc, char *argv[])
   int mpi=mpix*mpiy,flag;
   float *val[smax],*merge_val[smax];
   for (s=0;s<smax;s++){
-    val[s]=(float*)malloc(sizeof(double)*nx*ny);
-    merge_val[s]=(float*)malloc(sizeof(double)*nxall*nyall);
+    val[s]=(float*)malloc(sizeof(float)*nx*ny);
+    merge_val[s]=(float*)malloc(sizeof(float)*nxall*nyall);
   }
   
   for (n=nstt;n<nt;n++){
